Refactor component imports and add new components
