+++ app/_components/Hero.tsx
