+++ app/_components/ProjectsSection.tsx
