--- app/background/page.tsx
@@ -1,8 +1,0 @@
-import { ContactForm } from "@/components/contact-form";
-import React from "react";
-
-const page = () => {
-  return <ContactForm />;
-};
-
-export default page;
