+++ app/layout.tsx
@@ -1,11 +1,9 @@
-import type { Metadata } from "next";
-import { Inter, Poppins } from "next/font/google";
-import { Space_Grotesk } from "next/font/google";
-import "./globals.css";
 import { ThemeProvider } from "@/components/theme-provider";
-import { cn } from "@/lib/utils";
-import NavBar from "./components/NavBar";
-import Footer from "./components/Footer";
+import type { Metadata } from "next";
+import { Inter, Space_Grotesk } from "next/font/google";
+import Footer from "./_components/Footer";
+import NavBar from "./_components/NavBar";
+import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 const spaceGrotesk = Space_Grotesk({
@@ -30,7 +28,6 @@ export default function RootLayout({
         <ThemeProvider
           attribute="class"
           defaultTheme="dark"
-          // enableSystem
           disableTransitionOnChange
         >
           <NavBar />
